product_crud/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-crispy-forms==2.3
django-environ==0.12.0
django-extensions==3.2.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
ollama==0.4.7
pillow==11.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
twilio==9.4.5
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
